requirements.txt
openai>=1.40.0
pandas==2.0.3
numpy==1.26.4
pyyaml>=6.0.1
typer>=0.12.5
tqdm>=4.66.4
jinja2>=3.1.4
rich>=13.7.1
tenacity>=8.2.3
